Random subsample from the data of Angrist & Evans (1991).

Random subsample from the data of Angrist & Evans (1991).

Usage

AE98

Format

A data frame with 5,000 rows and 13 variables.

worked

Indicator equal to 1 if the mother is employed.

weeksw

Number of weeks of employment.

hoursw

Hours worked per week.

morekids

Indicator equal to 1 if the mother has more than 2 kids.

samesex

Indicator equal to 1 if the first two children are of the same sex.

age

Age in years.

agefst

Age in years at birth of the first child.

black

Indicator equal to 1 if the mother is black.

hisp

Indicator equal to 1 if the mother is Hispanic.

othrace

Indicator equal to 1 if the mother is neither black nor Hispanic.

educ

Years of education.

boy1st

Indicator equal to 1 if the first child is male.

boy2nd

Indicator equal to 1 if the second child is male.

Source

https://dataverse.harvard.edu/dataset.xhtml?persistentId=hdl:1902.1/11288