Estimator of the Average Treatment Effect.

Estimator of the average treatment effect.

Usage

ddml_ate(
  y,
  D,
  X,
  learners,
  learners_DX = learners,
  sample_folds = 2,
  ensemble_type = "nnls",
  shortstack = FALSE,
  cv_folds = 5,
  subsamples_D0 = NULL,
  subsamples_D1 = NULL,
  cv_subsamples_list_D0 = NULL,
  cv_subsamples_list_D1 = NULL,
  silent = F
)

Arguments

y

The outcome variable.

D

Binary endogenous variable of interest.

X

A (sparse) matrix of control variables.

learners

May take one of two forms, depending on whether a single learner or stacking with multiple learners is used for estimation of the conditional expectation functions. If a single learner is used, learners is a list with two named elements:

what The base learner function. The function must be such that it predicts a named input y using a named input X.

args Optional arguments to be passed to what.

If stacking with multiple learners is used, learners is a list of lists, each containing four named elements:

fun The base learner function. The function must be such that it predicts a named input y using a named input X.

args Optional arguments to be passed to fun.

assign_X An optional vector of column indices corresponding to control variables in X that are passed to the base learner.

Omission of the args element results in default arguments being used in fun. Omission of assign_X results in inclusion of all variables in X.

learners_DX

Optional argument to allow for different estimators of \(E[D|X]\). Setup is identical to learners.

sample_folds

Number of cross-fitting folds.

ensemble_type

Ensemble method to combine base learners into final estimate of the conditional expectation functions. Possible values are:

"nnls" Non-negative least squares.

"nnls1" Non-negative least squares with the constraint that all weights sum to one.

"singlebest" Select base learner with minimum MSPE.

"ols" Ordinary least squares.

"average" Simple average over base learners.

Multiple ensemble types may be passed as a vector of strings.

shortstack

Boolean to use short-stacking.

cv_folds

Number of folds used for cross-validation in ensemble construction.

subsamples_D0, subsamples_D1

List of vectors with sample indices for cross-fitting, corresponding to untreated and treated observations, respectively.

cv_subsamples_list_D0, cv_subsamples_list_D1

List of lists, each corresponding to a subsample containing vectors with subsample indices for cross-validation. Arguments are separated for untreated and treated observations, respectively.

silent

Boolean to silence estimation updates.

Value

ddml_ate returns an object of S3 class ddml_ate. An object of class ddml_ate is a list containing the following components:

ate

A vector with the average treatment effect estimates.

weights

A list of matrices, providing the weight assigned to each base learner (in chronological order) by the ensemble procedure.

mspe

A list of matrices, providing the MSPE of each base learner (in chronological order) computed by the cross-validation step in the ensemble construction.

psi_a, psi_b

Matrices needed for the computation of scores. Used in summary.ddml_ate().

learners,learners_DX, subsamples_D0,subsamples_D1, cv_subsamples_list_D0,cv_subsamples_list_D1, ensemble_type

Pass-through of selected user-provided arguments. See above.

Details

ddml_ate provides a double/debiased machine learning estimator for the average treatment effect in the interactive model given by

\(Y = g_0(D, X) + U,\)

where \((Y, D, X, U)\) is a random vector such that \(\operatorname{supp} D = \{0,1\}\), \(E[U\vert D, X] = 0\), and \(\Pr(D=1\vert X) \in (0, 1)\) with probability 1, and \(g_0\) is an unknown nuisance function.

In this model, the average treatment effect is defined as

\(\theta_0^{\textrm{ATE}} \equiv E[g_0(1, X) - g_0(0, X)]\).

References

Ahrens A, Hansen C B, Schaffer M E, Wiemann T (2023). "ddml: Double/debiased machine learning in Stata." https://arxiv.org/abs/2301.09397

Chernozhukov V, Chetverikov D, Demirer M, Duflo E, Hansen C B, Newey W, Robins J (2018). "Double/debiased machine learning for treatment and structural parameters." The Econometrics Journal, 21(1), C1-C68.

Wolpert D H (1992). "Stacked generalization." Neural Networks, 5(2), 241-259.

See also

Examples

# Construct variables from the included Angrist & Evans (1998) data
y = AE98[, "worked"]
D = AE98[, "morekids"]
X = AE98[, c("age","agefst","black","hisp","othrace","educ")]

# Estimate the average treatment effect using a single base learner, ridge.
ate_fit <- ddml_ate(y, D, X,
                    learners = list(what = mdl_glmnet,
                                    args = list(alpha = 0)),
                    sample_folds = 2,
                    silent = TRUE)
summary(ate_fit)
#> ATE estimation results: 
#>  
#>     Estimate Std. Error  t value    Pr(>|t|)
#>   -0.1513902 0.01567861 -9.65584 4.64339e-22

# Estimate the average treatment effect using short-stacking with base
#     learners ols, lasso, and ridge.
ate_fit <- ddml_ate(y, D, X,
                    learners = list(list(fun = ols),
                                    list(fun = mdl_glmnet),
                                    list(fun = mdl_glmnet,
                                         args = list(alpha = 0))),
                    ensemble_type = 'nnls',
                    shortstack = TRUE,
                    sample_folds = 2,
                    silent = TRUE)
summary(ate_fit)
#> ATE estimation results: 
#>  
#>        Estimate Std. Error   t value     Pr(>|t|)
#> nnls -0.1478009 0.01524395 -9.695709 3.144469e-22