Estimator for the Partially Linear Model.

Estimator for the partially linear model.

Usage

ddml_plm(
  y,
  D,
  X,
  learners,
  learners_DX = learners,
  sample_folds = 2,
  ensemble_type = "nnls",
  shortstack = FALSE,
  cv_folds = 5,
  subsamples = NULL,
  cv_subsamples_list = NULL,
  silent = FALSE
)

Arguments

y

The outcome variable.

D

The endogenous variable.

X

A (sparse) matrix of control variables.

learners

May take one of two forms, depending on whether a single learner or stacking with multiple learners is used for estimation of the conditional expectation function(s) \(E[Y|X]\) (and \(E[D|X]\) if learners_DX is not specified). If a single learner is used, learners is a list with two named elements:

what The base learner function. The function must be such that it predicts a named input y using a named input X.

args Optional arguments to be passed to what.

If stacking with multiple learners is used, learners is a list of lists, each containing four named elements:

fun The base learner function. The function must be such that it predicts a named input y using a named input X.

args Optional arguments to be passed to fun.

assign_X An optional vector of indices corresponding to features in X that are passed to the corresponding base learner.

learners_DX

Optional argument to allow for different estimators of \(E[D|X]\). Setup is identical to learners.

sample_folds

Number of cross-fitting folds.

ensemble_type

Ensemble method to combine base learners into final estimate of the conditional expectation functions. Possible values are:

"nnls" Non-negative least squares.

"nnls1" Non-negative least squares with the constraint that all weights sum to one.

"singlebest" Select base learner with minimum MSPE.

"ols" Ordinary least squares.

"average" Simple average over base learners.

Multiple ensemble types may be passed as a vector of strings.

shortstack

Boolean to use short-stacking.

cv_folds

Number of folds used for cross-validation in ensemble construction.

subsamples

List of vectors with sample indices for cross-fitting.

cv_subsamples_list

List of lists, each corresponding to a subsample containing vectors with vectors subsample indices for cross-validation.

silent

Boolean to silence estimation updates.

Value

ddml_plm returns an object of S3 class ddml_plm. An object of class ddml_plm is a list containing the following components:

coef

A vector with the \(\theta_0\) estimates.

weights

A list of matrices, providing the weight assigned to each base learner (in chronological order) by the ensemble procedure.

mspe

A list of matrices, providing the MSPE of each base learner (in chronological order) computed by the cross-validation step in the ensemble construction.

ols_fit

Object of class lm from the second stage regression of \(Y - \hat{E}[Y|X]\) on \(D - \hat{E}[D|X]\).

learners,learners_DX,subsamples, cv_subsamples_list,ensemble_type

Pass-through of selected user-provided arguments. See above.

Details

ddml_plm provides a double/debiased machine learning estimator for the parameter of interest \(\theta_0\) in the partially linear model given by

\(Y = \theta_0D + g_0(X) + U,\)

where \((Y, D, X, U)\) is a random vector such that \(E[Cov(U, D\vert X)] = 0\) and \(E[Var(D\vert X)] \neq 0\), and \(g_0\) is an unknown nuisance function.

See also

Examples

# Construct data from the included BLP_1995 data
y <- log(BLP_1995$share) - log(BLP_1995$outshr)
D <- BLP_1995$price
X <- as.matrix(subset(BLP_1995, select = c(air, hpwt, mpd, mpg, space)))
# Estimate the partially linear model using a single base learners: ridge.
plm_fit <- ddml_plm(y, D, X,
                    learners = list(what = mdl_glmnet,
                                    args = list(alpha = 0)),
                    sample_folds = 5,
                    silent = TRUE)
plm_fit$coef
#>         D_r 
#> -0.08840934